pyproject.toml
[tool.poetry]
name = "nlu"
version = "0.1.0"
description = ""
authors = ["Polyxya org"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
torch = { version = "=1.13.1", source = "torch" }
numpy = "^1.24.2"
uvicorn = "^0.21.0"
fastapi = "^0.94.1"
openai = "^0.27.2"
python-dotenv = "^1.0.0"
sentence-transformers = "^2.2.2"
surrealdb = "^0.3.0"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"
secondary = true
default = false


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
pre-commit = "^3.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
